[#5] 피드백 반영
* UserMapper.xml
- 쿼리 가독성 향상을 위해 개행추가

* FileService 인터페이스 추가
* 기존 FileService 클래스는 위의 인터페이스를 구현하고
  LocalFileService로 변경
- 어디에 저장하느냐에 따라 확장가능성이 있기 때문에 Controller가
  인터페이스에 의존하도록 해서 OCD조건 충족

* UserUpdateController 추가
- 기존 UserController로부터 유저 비밀번호/정보/프로필 사진에 대한 요청을
  받는 부분 책임 분리

* ControllerExceptionHandler 추가
- Controller 내부의 메소드 userProfileImageUpdate에서 일일이 try/catch를
  통해 예외를 잡았던 부분을 ExceptionHandler를 사용해 외부로 분리시킴

- 성격이 비슷한 두예외- FileNotUploadedException,
  FileNotDeletedException - 가 FileException을 상속받게 해서 공통 예외를
  만들고 이를  ExceptionHandler 내의 매개변수를 이용해서
  ExceptionHandler가 두 예외를 한번에 처리할 수 있도록 함

* Multipart File 최대 Size 증가
- 좀더 고화질의 사진업로드 시 용량 고려

* 테스트 코드
- 불필요하게 담겨 있는 코드/주석이 없는지 검토 후 수정 혹은 삭제
- 책임이전에 따라 테스트 클래스도 분리
(UserControllerTest -> UserControllerTest / UserUpdateControllerTest
