+++ src/main/java/me/soo/helloworld/exception/file/FileException.java
@@ -1,0 +1,12 @@
+package me.soo.helloworld.exception.file;
+
+public class FileException extends RuntimeException {
+
+    public FileException(String message) {
+        super(message);
+    }
+
+    public FileException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
