+++ src/main/java/me/soo/helloworld/exception/file/FileNotUploadedException.java
@@ -1,6 +1,6 @@
 package me.soo.helloworld.exception.file;
 
-public class FileNotUploadedException extends RuntimeException {
+public class FileNotUploadedException extends FileException {
 
     public FileNotUploadedException(String message, Throwable cause) {
         super(message, cause);
