+++ src/main/java/me/soo/helloworld/service/UserService.java
@@ -1,7 +1,6 @@
 package me.soo.helloworld.service;
 
 import lombok.RequiredArgsConstructor;
-import me.soo.helloworld.exception.file.FileNotUploadedException;
 import me.soo.helloworld.exception.IncorrectUserInfoException;
 import me.soo.helloworld.model.file.FileData;
 import me.soo.helloworld.model.user.*;
@@ -9,8 +8,6 @@ import me.soo.helloworld.repository.UserRepository;
 import me.soo.helloworld.util.encoder.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
 
 @Service
 @RequiredArgsConstructor
@@ -21,6 +18,8 @@ public class UserService {
     private final FileService fileService;
 
     private final PasswordEncoder passwordEncoder;
+
+    private final LoginService loginService;
 
     public void userSignUp(User user) {
         String encodedPassword = passwordEncoder.encode(user.getPassword());
@@ -50,6 +49,7 @@ public class UserService {
     public void userPasswordUpdate(String currentUserId, UserPasswordRequest userPasswordRequest) {
         String encodedPassword = passwordEncoder.encode(userPasswordRequest.getNewPassword());
         userRepository.updateUserPassword(currentUserId, encodedPassword);
+        loginService.logout();
     }
 
     public void userInfoUpdate(String userId, UserUpdateRequest updateRequest) {
