+++ src/main/java/me/soo/helloworld/util/http/HttpResponses.java
@@ -17,6 +17,4 @@ public class HttpResponses {
 
     public static final ResponseEntity<Void> HTTP_RESPONSE_NOT_FOUND = new ResponseEntity<>(HttpStatus.NOT_FOUND);
 
-    public static final ResponseEntity<Void> HTTP_RESPONSE_INTERNAL_SERVER_ERROR
-            = new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
 }
