+++ src/main/resources/application.properties
@@ -9,10 +9,10 @@ spring.servlet.multipart.enabled=true
 
 # 업로드 가능한 최대 파일 크기와 총 요청 크기설정 (단위 크기는 KB, MB 등을 숫자와 함께 지정가능하며 DataSize 클래스가 값을 변환하여 반환)
 # 조건을 초과한 요청이 들어오면 MultipartException 을 발생시킨다.
-spring.servlet.multipart.max-file-size=2MB
-spring.servlet.multipart.max-request-size=2MB
+spring.servlet.multipart.max-file-size=10MB
+spring.servlet.multipart.max-request-size=10MB
 
 # 임시파일을 만드는 기준: 메모리에서 바로 스트림으로 전달되는 크기
-#spring.servlet.multipart.file-size-threshold=
+spring.servlet.multipart.file-size-threshold=1MB
 
 file.upload.directory=D:\\Project\\Hello-World\\
