+++ src/main/resources/mappers/UserMapper.xml
@@ -3,41 +3,60 @@
 
 <mapper namespace="me.soo.helloworld.mapper.UserMapper">
     <insert id="insertUser" parameterType="me.soo.helloworld.model.user.User">
+
         INSERT INTO users(userId, password, email, gender, birthday, originCountry, livingCountry, livingTown, aboutMe,
                             profileImageName, profileImagePath)
+
         VALUES(#{userId}, #{password}, #{email}, #{gender}, #{birthday}, #{originCountry}, #{livingCountry}, #{livingTown},
                #{aboutMe}, #{profileImageName}, #{profileImagePath})
+
     </insert>
 
     <select id="isUserIdDuplicate" parameterType="String" resultType="boolean">
+
         SELECT EXISTS
+
             (SELECT userId FROM users WHERE userId = #{userId})
+
     </select>
 
     <select id="getUserById"
             parameterType="String" resultType="me.soo.helloworld.model.user.User">
+
         SELECT userId, password, email, gender, birthDay, originCountry, livingCountry, livingTown, aboutMe,
-               profileImageName, profileImagePath FROM users WHERE userId = #{userId}
+               profileImageName, profileImagePath
+
+        FROM users WHERE userId = #{userId}
+
     </select>
 
     <select id="getUserProfileImageById"
             parameterType="String" resultType="me.soo.helloworld.model.file.FileData">
+
         SELECT profileImageName, profileImagePath FROM users WHERE userId = #{userId}
+
     </select>
 
     <update id="updateUserPassword" parameterType="map">
+
         UPDATE users SET password = #{password} WHERE userId = #{userId}
+
     </update>
 
     <update id="updateUserInfo" parameterType="map">
+
         UPDATE users SET gender = #{updateRequest.gender}, livingCountry = #{updateRequest.livingCountry},
                          livingTown = #{updateRequest.livingTown}, aboutMe = #{updateRequest.aboutMe}
             WHERE userId = #{userId}
+
     </update>
 
     <update id="updateUserProfileImage" parameterType="map">
+
         UPDATE users SET profileImageName = #{profileImage.fileName}, profileImagePath = #{profileImage.filePath}
-        WHERE userId = #{userId}
+
+            WHERE userId = #{userId}
+
     </update>
 
 </mapper>
